Fix windows compilation errors
