+++ cpp/arcticdb/codec/segment.hpp
@@ -7,7 +7,7 @@
 
 #pragma once
 
-#include "util/buffer.hpp"
+#include <arcticdb/util/buffer.hpp>
 #include <arcticdb/storage/common.hpp>
 #include <google/protobuf/arena.h>
 #include <arcticdb/entity/field_collection.hpp>
