+++ cpp/arcticdb/entity/field_collection.hpp
@@ -8,9 +8,9 @@
 #pragma once
 
 #include <arcticdb/column_store/chunked_buffer.hpp>
-#include <util/cursored_buffer.hpp>
-#include <util/buffer.hpp>
-#include <entity/types.hpp>
+#include <arcticdb/util/cursored_buffer.hpp>
+#include <arcticdb/util/buffer.hpp>
+#include <arcticdb/entity/types.hpp>
 #include <arcticdb/column_store/column_data.hpp>
 
 namespace arcticdb {
