+++ cpp/arcticdb/pipeline/frame_utils.hpp
@@ -16,7 +16,7 @@
 #include <arcticdb/python/python_to_tensor_frame.hpp>
 #include <arcticdb/pipeline/string_pool_utils.hpp>
 #include <arcticdb/util/offset_string.hpp>
-#include <util/flatten_utils.hpp>
+#include <arcticdb/util/flatten_utils.hpp>
 #include <arcticdb/entity/timeseries_descriptor.hpp>
 #include <arcticdb/entity/type_utils.hpp>
