+++ cpp/arcticdb/stream/test/stream_test_common.cpp
@@ -5,10 +5,10 @@
  * As of the Change Date specified in that file, in accordance with the Business Source License, use of this software will be governed by the Apache License, version 2.0.
  */
 
-#include <arcticdb/stream/test/stream_test_common.hpp>
+#include <arcticdb/entity/types.hpp>
 #include <arcticdb/stream/piloted_clock.hpp>
 namespace arcticdb {
 
-std::atomic<timestamp> PilotedClock::time_{0};
+std::atomic<entity::timestamp> PilotedClock::time_{0};
 
 } //namespace arcticdb
