+++ cpp/arcticdb/version/test/test_merge.cpp
@@ -12,6 +12,8 @@
 #include <arcticdb/stream/aggregator.hpp>
 #include <arcticdb/util/test/generators.hpp>
 #include <arcticdb/stream/test/stream_test_common.hpp>
+#include <arcticdb/stream/test/test_store_common.hpp>
+
 
 struct MergeReadsTestStore : arcticdb::TestStore {
 protected:
